+++ include/boost/numeric/ublas/tensor/strides.hpp
@@ -8,6 +8,7 @@
 //  The authors gratefully acknowledge the support of
 //  Fraunhofer IOSB, Ettlingen Germany
 //
+/// \file strides.hpp Definition for the basic_strides template class
 
 
 #ifndef _BOOST_UBLAS_TENSOR_STRIDES_
@@ -177,6 +178,50 @@ protected:
 
 template<class layout_type>
 using strides = basic_strides<std::size_t, layout_type>;
+
+namespace detail {
+
+
+/** @brief Memory access function with multi-indices
+ *
+ * @code auto m = access(0, 3,4,5); @endcode
+ *
+ * @param[in] i multi-index vector of length p
+ * @returns relative memory location depending on \c i
+*/
+BOOST_UBLAS_INLINE
+template<class size_type, class layout_type>
+auto access(std::vector<size_type> const& i, basic_strides<size_type,layout_type> const& w)
+{
+	const auto p = i.size();
+	size_type sum = 0u;
+	for(auto r = 0u; r < p; ++r)
+		sum += i[r]*w[r];
+	return sum;
+}
+
+/** @brief Memory access function with multi-indices
+ *
+ * @code auto m = access(0, 3,4,5); @endcode
+ *
+ *
+ * @param[in] i   first element of the partial multi-index
+ * @param[in] is  the following elements of the partial multi-index
+ * @param[in] sum the current
+ * @returns relative memory location depending on \c i
+*/
+BOOST_UBLAS_INLINE
+template<std::size_t r, class layout_type, class size_type, class ... size_types>
+auto access(size_type sum, basic_strides<size_type, layout_type> const& w, size_type i, size_types ... is)
+{
+	sum+=i*w[r];
+	if constexpr (sizeof...(is) == 0)
+		return sum;
+	else
+		return detail::access<r+1>(sum,w,std::forward<size_type>(is)...);
+}
+
+}
 
 }}}