+++ test/tensor/test_strides.cpp
@@ -3,6 +3,13 @@
 //  Distributed under the Boost Software License, Version 1.0. (See
 //  accompanying file LICENSE_1_0.txt or copy at
 //  http://www.boost.org/LICENSE_1_0.txt)
+//
+//  The authors gratefully acknowledge the support of
+//  Fraunhofer IOSB in producing this work.
+//
+//  And we acknowledge the support from all contributors.
+
+
 
 #include <boost/test/unit_test.hpp>
 #include <boost/numeric/ublas/tensor/strides.hpp>
@@ -38,7 +45,7 @@ BOOST_AUTO_TEST_CASE_TEMPLATE( test_strides_ctor, value, test_types)
 {
 	using namespace boost::numeric;
 
-	using extents_type  = ublas::extents;
+	using extents_type  = ublas::basic_extents<unsigned>;
 	using strides_type = ublas::strides<value>;
 
 	strides_type         s0{};
@@ -80,7 +87,7 @@ BOOST_AUTO_TEST_CASE( test_strides_ctor_access_first_order)
 {
 	using namespace boost::numeric;
 
-	using extents_type  = ublas::extents;
+	using extents_type  = ublas::basic_extents<unsigned>;
 	using strides_type =  ublas::strides<ublas::first_order>;
 
 	strides_type         s1{extents_type{1,1}};
@@ -132,7 +139,7 @@ BOOST_AUTO_TEST_CASE( test_strides_ctor_access_last_order)
 {
 	using namespace boost::numeric;
 
-	using extents_type  = ublas::extents;
+	using extents_type  = ublas::basic_extents<unsigned>;
 	using strides_type =  ublas::strides<ublas::last_order>;
 
 	strides_type         s1{extents_type{1,1}};