+++ test/tensor/test_tensor.cpp
@@ -3,65 +3,27 @@
 //  Distributed under the Boost Software License, Version 1.0. (See
 //  accompanying file LICENSE_1_0.txt or copy at
 //  http://www.boost.org/LICENSE_1_0.txt)
+//
+//  The authors gratefully acknowledge the support of
+//  Fraunhofer IOSB in producing this work.
+//
+//  And we acknowledge the support from all contributors.
+
+
 
 #include <random>
-#include <boost/numeric/ublas/tensor.hpp>
+#include <boost/numeric/ublas/tensor/tensor.hpp>
 
 #define BOOST_TEST_DYN_LINK
 #define BOOST_TEST_MODULE TestTensor
+
 #include <boost/test/unit_test.hpp>
+#include "utility.hpp"
 
 BOOST_AUTO_TEST_SUITE ( test_tensor, * boost::unit_test::depends_on("test_extents") ) ;
 
 
-
-
-
-template<class ... types>
-struct zip_helper;
-
-template<class type1, class ... types3>
-struct zip_helper<std::tuple<types3...>, type1>
-{
-	template<class ... types2>
-	struct with
-	{
-		using type = std::tuple<types3...,std::pair<type1,types2>...>;
-	};
-	template<class ... types2>
-	using with_t = typename with<types2...>::type;
-};
-
-
-template<class type1, class ... types3, class ... types1>
-struct zip_helper<std::tuple<types3...>, type1, types1...>
-{
-	template<class ... types2>
-	struct with
-	{
-		using next_tuple = std::tuple<types3...,std::pair<type1,types2>...>;
-		using type       = typename zip_helper<next_tuple, types1...>::template with<types2...>::type;
-	};
-
-	template<class ... types2>
-	using with_t = typename with<types2...>::type;
-};
-
-template<class ... types>
-using zip = zip_helper<std::tuple<>,types...>;
-
 using test_types = zip<int,long,float,double>::with_t<boost::numeric::ublas::first_order, boost::numeric::ublas::last_order>;
-
-// creates e.g.
-//using test_types =
-//std::tuple<
-//std::pair<float, boost::numeric::ublas::first_order>,
-//std::pair<float, boost::numeric::ublas::last_order >,
-//std::pair<double,boost::numeric::ublas::first_order>,
-//std::pair<double,boost::numeric::ublas::last_order >
-//>;
-//static_assert(std::is_same< std::tuple_element_t<0,std::tuple_element_t<0,test_types2>>, float>::value,"should be float ");
-//static_assert(std::is_same< std::tuple_element_t<1,std::tuple_element_t<0,test_types2>>, boost::numeric::ublas::first_order>::value,"should be boost::numeric::ublas::first_order ");
 
 
 BOOST_AUTO_TEST_CASE_TEMPLATE( test_tensor_ctor, value,  test_types)
@@ -109,7 +71,7 @@ BOOST_AUTO_TEST_CASE_TEMPLATE( test_tensor_ctor, value,  test_types)
 
 
 struct fixture {
-	using extents_type = boost::numeric::ublas::extents;
+	using extents_type = boost::numeric::ublas::basic_extents<std::size_t>;
 	fixture() : extents{
 				extents_type{},    // 0
 				extents_type{1,1}, // 1
@@ -134,7 +96,7 @@ BOOST_FIXTURE_TEST_CASE_TEMPLATE( test_tensor_ctor_extents, value,  test_types, 
 	using layout_type = typename value::second_type;
 	using tensor_type = ublas::tensor<value_type, layout_type>;
 
-	auto check = [](ublas::extents const& e) {
+	auto check = [](auto const& e) {
 		auto t = tensor_type{e};
 		BOOST_CHECK_EQUAL (  t.size() , e.product() );
 		BOOST_CHECK_EQUAL (  t.rank() , e.size() );
@@ -160,7 +122,7 @@ BOOST_FIXTURE_TEST_CASE_TEMPLATE( test_tensor_copy_ctor, value,  test_types, fix
 	using layout_type = typename value::second_type;
 	using tensor_type = ublas::tensor<value_type, layout_type>;
 
-	auto check = [](ublas::extents const& e)
+	auto check = [](auto const& e)
 	{
 		auto r = tensor_type{e};
 		auto t = r;
@@ -194,7 +156,7 @@ BOOST_FIXTURE_TEST_CASE_TEMPLATE( test_tensor_copy_move_ctor, value,  test_types
 	using layout_type = typename value::second_type;
 	using tensor_type = ublas::tensor<value_type, layout_type>;
 
-	auto check = [](ublas::extents const& e)
+	auto check = [](auto const& e)
 	{
 		auto r = tensor_type{e};
 		auto t = std::move(r);
@@ -230,7 +192,7 @@ BOOST_FIXTURE_TEST_CASE_TEMPLATE( test_tensor_ctor_extents_init, value,  test_ty
 	using distribution_type = std::conditional_t<std::is_integral_v<value_type>, std::uniform_int_distribution<>, std::uniform_real_distribution<> >;
 	auto distribution = distribution_type(1,6);
 
-	auto check = [&distribution,&generator](ublas::extents const& e) {
+	auto check = [&distribution,&generator](auto const& e) {
 		auto r = static_cast<value_type>(distribution(generator));
 		auto t = tensor_type{e,r};
 		for(auto i = 0ul; i < t.size(); ++i)
@@ -251,7 +213,7 @@ BOOST_FIXTURE_TEST_CASE_TEMPLATE( test_tensor_ctor_extents_array, value,  test_t
 	using tensor_type = ublas::tensor<value_type, layout_type>;
 	using array_type  = typename tensor_type::array_type;
 
-	auto check = [](ublas::extents const& e) {
+	auto check = [](auto const& e) {
 		auto a = array_type(e.product());
 		std::iota(a.begin(), a.end(), value_type{});
 		auto t = tensor_type{e, a};
@@ -273,12 +235,15 @@ BOOST_FIXTURE_TEST_CASE_TEMPLATE( test_tensor_read_write_single_index_access, va
 	using layout_type = typename value::second_type;
 	using tensor_type = ublas::tensor<value_type, layout_type>;
 
-	auto check = [](ublas::extents const& e) {
+	auto check = [](auto const& e) {
 		auto t = tensor_type{e};
 		auto v = value_type {};
 		for(auto i = 0ul; i < t.size(); ++i, ++v){
 			t[i] = v;
 			BOOST_CHECK_EQUAL( t[i], v );
+
+			t(i) = v;
+			BOOST_CHECK_EQUAL( t(i), v );
 		}
 	};
 
@@ -340,7 +305,7 @@ BOOST_FIXTURE_TEST_CASE_TEMPLATE( test_tensor_read_write_multi_index_access_at, 
 						BOOST_CHECK_EQUAL(t.at(k[0],k[1],k[2],k[3]), v++);
 	};
 
-	auto check = [check1,check2,check3,check4](ublas::extents const& e) {
+	auto check = [check1,check2,check3,check4](auto const& e) {
 		auto t = tensor_type{e};
 		auto v = value_type {};
 		for(auto i = 0ul; i < t.size(); ++i)
@@ -367,7 +332,7 @@ BOOST_FIXTURE_TEST_CASE_TEMPLATE( test_tensor_reshape, value,  test_types, fixtu
 	using layout_type = typename value::second_type;
 	using tensor_type = ublas::tensor<value_type, layout_type>;
 
-	auto check = [](ublas::extents const& efrom, ublas::extents const& eto)
+	auto check = [](auto const& efrom, auto const& eto)
 	{
 		auto v = value_type {};
 		++v;
@@ -402,7 +367,7 @@ BOOST_FIXTURE_TEST_CASE_TEMPLATE( test_tensor_swap, value,  test_types, fixture)
 	using layout_type = typename value::second_type;
 	using tensor_type = ublas::tensor<value_type, layout_type>;
 
-	auto check = [](ublas::extents const& e_t, ublas::extents const& e_r)
+	auto check = [](auto const& e_t, auto const& e_r)
 	{
 		auto v = value_type {} + 1;
 		auto w = value_type {} + 2;
@@ -444,7 +409,7 @@ BOOST_FIXTURE_TEST_CASE_TEMPLATE( test_tensor_standard_iterator, value,  test_ty
 	using iterator_type = typename tensor_type::iterator;
 	using const_iterator_type = typename tensor_type::const_iterator;
 
-	auto check = [](ublas::extents const& e)
+	auto check = [](auto const& e)
 	{
 		auto v = value_type {} + 1;
 		auto t = tensor_type{e, v};